requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pycryptodome==3.19.0 